+++ jupytext/cell_reader.py
@@ -58,6 +58,15 @@ def count_lines_to_next_cell(cell_end_marker, next_cell_start,
     return 1
 
 
+def last_two_lines_blank(source):
+    """Are the two last lines blank, and not the third last one?"""
+    if len(source) < 3:
+        return False
+    return (not _BLANK_LINE.match(source[-3]) and
+            _BLANK_LINE.match(source[-2]) and
+            _BLANK_LINE.match(source[-1]))
+
+
 class CellReader():
     """A class that can read notebook cells from their text representation"""
 
@@ -114,8 +123,8 @@ class CellReader():
             self.language, self.metadata = \
                 rmd_options_to_metadata('r ' + _CODE_OPTION_R.findall(line)[0])
 
-        if self.ext == '.py' and _CODE_OPTION_PY.match(line):
-            self.language = 'python'
+        if self.ext in ['.py', '.jl'] and _CODE_OPTION_PY.match(line):
+            self.language = 'python' if self.ext == '.py' else 'julia'
             self.metadata = json_options_to_metadata(
                 _CODE_OPTION_PY.match(line).group(3))
 
@@ -198,17 +207,10 @@ class CellReader():
         """Given that this is a code cell, return position of
         end of cell marker, and position of next cell start"""
         self.cell_type = 'code'
-        parser = StringParser('python' if self.ext == '.py' else 'R')
-        empty = True
+        parser = StringParser('python' if self.ext in ['.py', '.jl'] else 'R')
         for i, line in enumerate(lines):
             # skip cell header
             if self.metadata is not None and i == 0:
-                continue
-
-            # Read something!
-            if not _BLANK_LINE.match(line):
-                empty = False
-            elif empty:
                 continue
 
             if parser.is_quoted():
@@ -228,14 +230,17 @@ class CellReader():
                     return i, i + 1, True
             elif _BLANK_LINE.match(line):
                 if not next_code_is_indented(lines[i:]):
-                    return i, i + 1, False
+                    if i > 0:
+                        return i, i + 1, False
+                    if len(lines) == 1 or _BLANK_LINE.match(lines[1]):
+                        return 1, 2, False
 
         return len(lines), len(lines), False
 
     def find_cell_end(self, lines):
         """Return position of end of cell marker, and position
         of first line after cell"""
-        if self.ext == '.py':
+        if self.ext in ['.py', '.jl']:
             return self.find_cell_end_py(lines)
         if self.ext in ['.Rmd', '.md']:
             return self.find_cell_end_rmd(lines)
@@ -260,17 +265,23 @@ class CellReader():
         if self.cell_type == 'code':
             if self.ext != '.md' and is_active(self.ext, self.metadata):
                 unescape_magic(source, self.language or 'python')
-            elif self.ext in ['.py', '.R']:
+            elif self.ext in ['.py', '.jl', '.R']:
                 source = uncomment(source)
 
         if self.cell_type == 'code' or self.ext in ['.Rmd', '.md']:
             unescape_code_start(source, self.ext, self.language or
                                 ('python' if self.ext != '.Rmd' else None))
 
-        if self.cell_type == 'markdown' and self.ext in ['.py', '.R']:
+        if self.cell_type == 'markdown' and self.ext in ['.py', '.jl', '.R']:
             source = self.markdown_unescape(source)
 
         self.content = source
+
+        # Exactly two empty lines at the end?
+        if (self.ext == '.py' and explicit_eoc and
+                last_two_lines_blank(source)):
+            self.content = source[:-2]
+            self.metadata['lines_to_end_of_cell_marker'] = 2
 
         # Is this a raw cell?
         if not is_active('ipynb', self.metadata) or (
@@ -280,11 +291,16 @@ class CellReader():
                 del self.metadata['active']
             self.cell_type = 'raw'
 
-        # Does the next cell start one line later?
+        # Explicit end of cell marker?
         if (next_cell_start + 1 < len(lines) and
                 _BLANK_LINE.match(lines[next_cell_start]) and
                 not _BLANK_LINE.match(lines[next_cell_start + 1])):
             next_cell_start += 1
+        elif (explicit_eoc and next_cell_start + 2 < len(lines) and
+              _BLANK_LINE.match(lines[next_cell_start]) and
+              _BLANK_LINE.match(lines[next_cell_start + 1]) and
+              not _BLANK_LINE.match(lines[next_cell_start + 2])):
+            next_cell_start += 2
 
         self.lines_to_next_cell = count_lines_to_next_cell(
             cell_end_marker,