+++ jupytext/jupytext.py
@@ -23,7 +23,7 @@ from .languages import default_language_from_metadata_and_ext, \
 from .cell_to_text import CellExporter
 from .cell_reader import CellReader
 
-NOTEBOOK_EXTENSIONS = ['.ipynb', '.Rmd', '.md', '.py', '.R']
+NOTEBOOK_EXTENSIONS = ['.ipynb', '.Rmd', '.md', '.jl', '.py', '.R']
 
 
 def markdown_comment(ext):
@@ -77,11 +77,13 @@ class TextNotebookReader(NotebookReader):
 
         set_main_and_cell_language(metadata, cells, self.ext)
 
-        if 'nbrmd_formats' in metadata:
-            metadata['jupytext_formats'] = metadata.pop('nbrmd_formats')
+        # Backward compatibility with nbrmd
+        for key in ['nbrmd_formats', 'nbrmd_format_version']:
+            if key in metadata:
+                metadata[key.replace('nbrmd', 'jupytext')] = \
+                    metadata.pop(key)
 
-        notebook = new_notebook(cells=cells, metadata=metadata)
-        return notebook
+        return new_notebook(cells=cells, metadata=metadata)
 
 
 class TextNotebookWriter(NotebookWriter):
@@ -125,8 +127,12 @@ class TextNotebookWriter(NotebookWriter):
 
             # remove end of cell marker when redundant
             # with next explicit marker
-            if self.ext == '.py' and cell.is_code() and text[-1] == '# -':
-                if i + 1 >= len(texts) or \
+            if self.ext in ['.py', '.jl'] and cell.is_code() \
+                    and text[-1] == '# -':
+                if cell.lines_to_end_of_cell_marker:
+                    text = text[:-1] + \
+                           [''] * cell.lines_to_end_of_cell_marker + ['# -']
+                elif i + 1 >= len(texts) or \
                         (texts[i + 1][0].startswith('# + {')):
                     text = text[:-1]